Move CloudSdkProcessFacade to another bundle
