+++ plugins/com.google.cloud.tools.eclipse.appengine.deploy/src/com/google/cloud/tools/eclipse/appengine/deploy/messages.properties
@@ -6,7 +6,6 @@ task.name.publish.war=Preparing deploy artifact
 deploy.job.name=Deploying to App Engine
 deploy.job.staging.failed=Staging failed. Check the error message in the Console View.
 war.publishing.failed=war publishing failed.
-cloudsdk.process.failed=Process exited with error code {0}
 save.credential.failed=Error temporarily saving credential
 browser.launch.failed=Error launching deployed app in browser
 browser.launch.title=App Engine Deploy - {0}
