+++ plugins/com.google.cloud.tools.eclipse.sdk/META-INF/MANIFEST.MF
@@ -11,9 +11,12 @@ Require-Bundle: com.google.cloud.tools.appengine;bundle-version="0.3.6",
  org.eclipse.equinox.preferences;bundle-version="3.5.300",
  org.eclipse.e4.core.contexts;bundle-version="1.4.0",
  org.eclipse.e4.core.di;bundle-version="1.5.0"
-Import-Package: com.google.common.annotations;version="[20.0.0,21.0.0)",
- com.google.common.collect;version="[20.0.0,21.0.0)",
+Import-Package: com.google.cloud.tools.eclipse.util,
+ com.google.cloud.tools.eclipse.util.status,
+ com.google.common.annotations;version="[20.0.0,21.0.0)",
  com.google.common.base;version="[20.0.0,21.0.0)",
+ com.google.common.collect;version="[20.0.0,21.0.0)",
+ org.eclipse.core.runtime;version="3.5.0",
  org.eclipse.jface.preference,
  org.eclipse.ui.preferences,
  org.osgi.framework;version="1.8.0",
