+++ plugins/com.google.cloud.tools.eclipse.sdk/src/com/google/cloud/tools/eclipse/sdk/messages.properties
@@ -1,0 +1,1 @@
+cloudsdk.process.failed=Process exited with error code {0}
